pyproject.toml
[project]
name = "simulador-proposta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.3",
    "matplotlib>=3.10.3",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
    "streamlit-aggrid>=1.1.6",
]
